requirements.txt
bcrypt
pillow
python-jose
twilio
yenot
